+++ src/blocks/u64_blocks.rs
@@ -12,8 +12,7 @@ pub struct Block {
     runends: u64,
     counts: u64,
     remainders: [Remainder; SLOTS_PER_BLOCK],
-    offset: u16,
-    // padding: [u16; 3],
+    offset: u64,
 }
 
 pub struct U64Blocks {
